requirements.txt
numpy==1.26.2
tokenizers==0.19.1
python-dotenv==1.0.0
soundfile==0.12.1
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.9
Pillow==10.1.0